+++ sub/ext/ext_test.go
@@ -18,7 +18,7 @@ func (s *ExtSuite) SetupTest() {
 
 func (s *ExtSuite) TestNew() {
 	v := validator.New([]string{}, []string{}, "TRACE")
-	exts := New(v)
+	exts := New(v, nil)
 
 	s.IsType(&root{}, exts["root"])
 }