+++ sub/handler.go
@@ -6,6 +6,7 @@ import (
 	"net/http"
 	"strings"
 
+	"github.com/mattevans/postmark-go"
 	"github.com/microcosm-cc/bluemonday"
 	"maunium.net/go/mautrix"
 	"maunium.net/go/mautrix/id"
@@ -20,6 +21,11 @@ import (
 type Sender interface {
 	Send(id.RoomID, string)
 	SendFile(id.RoomID, *mautrix.ReqUploadMedia)
+}
+
+// EmailSender interface
+type EmailSender interface {
+	Send(*postmark.Email) (*postmark.EmailResponse, *postmark.Response, error)
 }
 
 // Handler is an HTTP forms handler
@@ -37,7 +43,7 @@ type Handler struct {
 const redirect = "<html><head><title>Redirecting...</title><meta http-equiv=\"Refresh\" content=\"0; url='{{ .URL }}'\" /></head><body>Redirecting to <a href='{{ .URL }}'>{{ .URL }}</a>..."
 
 // NewHandler creates new HTTP forms handler
-func NewHandler(forms map[string]*config.Form, spam *config.Spam, sender Sender, loglevel string) *Handler {
+func NewHandler(forms map[string]*config.Form, spam *config.Spam, pm EmailSender, sender Sender, loglevel string) *Handler {
 	v := validator.New(spam.Hosts, spam.Emails, loglevel)
 	h := &Handler{
 		redirectTpl: template.Must(template.New("redirect").Parse(redirect)),
@@ -46,7 +52,7 @@ func NewHandler(forms map[string]*config.Form, spam *config.Spam, sender Sender,
 		forms:       forms,
 		spam:        spam,
 		log:         logger.New("sub.", loglevel),
-		ext:         ext.New(v),
+		ext:         ext.New(v, pm),
 		v:           v,
 	}